Move Histogram into its own mod
This makes room in lib.rs for new functions.
